requirements.txt
requests
json
python-dotenv

